requirements.txt
boto3==1.24.31
botocore==1.27.31
Django==3.2
django-cleanup==6.0.0
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
dj-database-url==0.5.0
gunicorn==20.1.0
pillow==9.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
yake==0.4.8

# Dev requirements
coverage==6.4.1
tblib==1.7.0
